Improvement to prevent regroot fetch at proxy and group creation
As UI will ask for repo root (first screen shown on UI after
creation is browse storage), that would kick in with complete
registry root fetch. This is ugly, as not only puts huge load
on server, but also inevitably timeouts and shows an error in
UI.

So, we step back a bit from spec (by serving metadata from path
"/") as npm CLI uses "/-/all" path for that anyway.

But this improves user experience by not fetching regroot and
not having an error in UI.

npm search is anyway not quite used feature of npm cli. So,
originally the fetch happened was maybe even unnecessary.