NEXUS-6882: Removal of the intent use
As it turns out, use of "intent" to improve performance
is not playing well with other load running in parallel.
Removing it as advised by Orient folks.

Also, returning the TX boundaries to be per-document, as
the regRoot update not only updates each document but also
has it's own "marker" document that preserves the state of
the operation, hence, no need for huge TX spawning ALL
of the records in the database. Due to changes happened
in "NEXUS-6863: Make failure propagate to the client" issue
the search result will never result in partial result.